Split markdown security selector into own file
